+++ CHANGELOG.md
@@ -17,6 +17,10 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+
+- `social-polls`
+  - [#3122](https://github.com/wix/ricos/pull/3122) Polls a11y
 
 </details>
 <hr/>
@@ -27,6 +31,7 @@
 
 - `plugin-headings toolbars-new`
   - [#2987](https://github.com/wix/ricos/pull/2987) `allowHeadingCustomization` prop enables headings customization modal in the new formatting toolbar
+
 ### :bug: Bug Fix
 
 - `vertical-embed`